fix: use u32 for Discord embed color
We were using u8 for Discord embed colors, which isn't enough since
colors require 3-4 bytes. u32 is much better suited for this, let's use
it.

+++ src/discord.rs
@@ -51,7 +51,7 @@ pub struct Embed {
     pub thumbnail: Option<String>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub color: Option<u8>,
+    pub color: Option<u32>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     pub url: Option<String>,
